backend/requirements.txt
APScheduler==3.10.4
bcrypt==4.1.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
google-auth==2.34.0
gunicorn==22.0.0
marshmallow==3.21.3
python-dotenv==1.0.1
requests==2.32.3
